refactor(arch): relocate harvest to internal/modules/harvest (Phase 3 pilot 1a)
First concrete Phase 3 step (PHASE3_EXTRACTION_PLAN.md): establish the
internal/modules/ ring by relocating the harvest package, the lowest-risk
pilot (cold, single external importer, already I/O-light).

Pure mechanical move — `git mv internal/harvest internal/modules/harvest` (27
renames) plus the import-path rewrite in the three external importers
(api/modules.go, cmd/seed/cmd_serve.go, cmd_service_windows.go). No logic, no
behavior change: the package name stays `harvest`, only its import path moves.

Resliced from the plan's PR1 into a standalone relocation (1a) so it is atomic
and trivially reviewable; ports + adapters + depguard land next as 1b. The
module still imports internal/{database,config,health} directly for now — those
become ReportRepo / cfg.Harvest / HealthSource ports in 1b, before depguard
domain-purity turns on.

Verified: go build ./... clean, golden HTTP suite green (behavior gate),
harvest unit tests pass, gofumpt + lint 0 issues.
