+++ internal/modules/harvest/aggregator/internal_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/krisarmstrong/seed/internal/harvest/aggregator"
+	"github.com/krisarmstrong/seed/internal/modules/harvest/aggregator"
 )
 
 // ----------------------------------------------------------------------------
